requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
pycryptodome
